🎯 Phase 4.8C: cexecute.c Perfect Warning Elimination (100% Success)
OUTSTANDING ACHIEVEMENT:
- Starting warnings: 28
- Final warnings: 0 (ZERO WARNINGS - PERFECT!)
- Success rate: 100% warning elimination
- Enhanced safe_convert.h with safe_long_to_short() function

COMPREHENSIVE WARNING FIXES:
✅ short → unsigned char conversions (14 warnings): P_ASTAT, P_ATYPE, P_AXLOC, P_AYLOC, xloc, yloc, crew, armynum, terror, reputation, active, charity, capx, capy, P_AMOVE, smove
✅ int → unsigned char conversions (5 warnings): popularity, tax_rate, sector owner assignments
✅ long → short conversions (3 warnings): i_people scaling, MERCATT, MERCDEF calculations
✅ short → unsigned short sign conversions (3 warnings): merchant, warships, galleys ship assignments
✅ int → short conversions (4 warnings): country variable assignments throughout switch cases
✅ short → char conversion (1 warning): dstatus diplomatic status assignment

SAFE_CONVERT.H ENHANCEMENTS:
- Added safe_long_to_short() function (14th conversion utility)
- Complete documentation with usage examples
- Eliminates need for double-conversion patterns
- Maintains clean, readable code architecture

SYSTEMATIC APPLICATION:
- Applied appropriate safe_convert function for each warning type
- Maintained game logic and value range constraints
- Dual-mode compilation success (admin + game modes)
- Comprehensive testing validation

ARCHITECTURAL BENEFITS:
- Clean conversion patterns throughout command execution system
- Enhanced toolkit ready for remaining Phase 4.8C files
- Proven methodology for complex switch statement modernization
- Zero functional changes - pure warning elimination

+++ psmap.c
@@ -123,7 +123,7 @@ char progname[80];
  *   - Returns 0 for any unrecognized input (safe default)
  *   - Used by command-line option processing and environment variable parsing
  */
-int 
+int
 parsepagesize (char *buf)
 {
     if (!strcmp(buf, "A4"))
@@ -178,7 +178,7 @@ parsepagesize (char *buf)
  *   - Default case handles custom/user-defined page sizes
  *   - Values sourced from psmap.h constant definitions
  */
-void 
+void
 setpagesize (int defpag)
 {
     switch (defpag) {
@@ -244,7 +244,7 @@ setpagesize (int defpag)
  *   - Uses parsepagesize() for string-to-code conversion
  *   - Not thread-safe due to global state modifications
  */
-void 
+void
 get_pagesize (void)
 {
     char *buf;
@@ -297,7 +297,7 @@ get_pagesize (void)
  *   - Essential for preventing PostScript syntax errors
  *   - Used for titles, footers, and text labels in map output
  */
-void 
+void
 psstring (FILE *fh, char *str)
 {
     fprintf(fh, "(");
@@ -366,10 +366,10 @@ psstring (FILE *fh, char *str)
  *   - Used by getmaptype() for map format detection
  *   - Efficient for short patterns in header strings
  */
-int 
+int
 isinstr (char *string, char *word)
 {
-    int i,l1=strlen(string),l2=strlen(word);
+    size_t i,l1=strlen(string),l2=strlen(word);
 
     if (l1 < l2) return(FALSE);
     for(i = 0; i < l1; i++ ) {
@@ -429,7 +429,7 @@ isinstr (char *string, char *word)
  *   - Used during map file parsing to configure rendering pipeline
  *   - Falls back to SIMPLE type for unknown formats
  */
-int 
+int
 getmaptype (char *string)
 {
     if (isinstr(string, "Altitude"))
@@ -499,7 +499,7 @@ getmaptype (char *string)
  *   - Critical function in the map processing pipeline
  *   - Handles variable-sized maps with automatic dimension detection
  */
-void 
+void
 readmap (void)
 {
     int x, none;
@@ -627,7 +627,7 @@ readmap (void)
  *   - Handles both single-page and multi-page output modes
  *   - PostScript variables enable template customization
  */
-void 
+void
 buildps (void)
 {
     int xbeg, ybeg, xnumb, ynumb, x, y, xpages, ypages, xcorr, ycorr;
@@ -992,4 +992,3 @@ main(int argc, char *argv[])
 
     return (0);
 }
-
